tests/vm: Add configuration to basevm.py
Added use of a configuration to tests/vm/basevm.py.
This allows for providing alternate configurations to the vm being
created/launched. cpu, machine, memory, and numa configuration are all
examples of configuration which we might want to vary on the vm being created
or launched.
This will for example allow for creating an aarch64 vm.