accel/tcg:  interrupt/exception handling treats BQL as optional.
This change removes the implied BQL from the cpu_handle_interrupt,
and cpu_handle_exception paths. We can now select per-arch if
the BQL is needed or not.  This essentially leaves it up to the
arch to make the change to remove the BQL for this path when
it makes sense.  But by default, the core code holds the
BQL.