accel/tcg: Push BQL on ->cpu_exec_interrupt down into per-arch implementation
Pushing the BQL down into the per-arch implementation.  The first step is
to remove the BQL from cpu_handle_interrupt.
We then added the *_cpu_exec_interrupt functions which get the BQL and then
call to *_cpu_exec_interrupt_locked.  We also pointed the per-arch
->cpu_exec_interrupt at the new *cpu_exec_interrupt.

The purpose of this set of changes is to set the groundwork
so that an arch could move towards removing
the BQL from the cpu_handle_interrupt/exception paths.

This approach was suggested by Paolo Bonzini.
For reference, here are key posts in the discussion, explaining
the reasoning/benefits of this approach.
https://lists.gnu.org/archive/html/qemu-devel/2020-08/msg00784.html
https://lists.gnu.org/archive/html/qemu-devel/2020-07/msg08731.html
https://lists.gnu.org/archive/html/qemu-devel/2020-08/msg00044.html

Signed-off-by: Robert Foley <robert.foley@linaro.org>